+++ lib/actions/FileActions.js
@@ -1,6 +1,6 @@
 var FileActions = {};
 var Dispatcher = require('../distpatcher.js');
-var EVENTS = require('../constatns/events.js');
+var EVENTS = require('./events.js');
 
 
 FileActions.openDirectory = function(dir,config) {
