+++ lib/distpatcher.js
@@ -9,10 +9,6 @@ var Dispatcher = _.extend(new EventEmitter(), {
 
 
 Dispatcher.handleAction = function(action) {
-  // var payload = {
-  //   action: action
-  // };
   this.emit(ACTION_EVENT, action);
-  // this.dis
 }
 module.exports = Dispatcher;
