+++ lib/ipcManager.js
@@ -3,19 +3,12 @@ import {
   ipcMain
 } from 'electron';
 var FileStore = require('./stores/FileStore.js');
-var EVENTS = require('./constatns/events.js');
+var EVENTS = require('./constants/events.js');
 var FileActions = require('./actions/FileActions.js');
-// FileStore.on(EVENTS.OPEN_FILES, function() {
-//   FileStore.getFiles();
-// })
-
 
 ipcMain.on('clickDictionary', (event, dic) => {
   console.log('ipc on clickDictionary')
   FileActions.clickDictionary(dic);
-  // fileManager.openItem(file);
-
-  // event.sender.send('asynchronous-reply', 'pong');
 });
 
 ipcMain.on('clickFile', (event, file) => {
