+++ lib/mainWindowManagerFactory.js
@@ -1,6 +1,6 @@
 // send message to render procress
 var FileStore = require('./stores/FileStore.js');
-var EVENTS = require('./constatns/events.js');
+var EVENTS = require('./constants/events.js');
 
 function mainWindowManagerFactory(mainWindow) {
   let mainWindowManager = {};
