+++ lib/stores/FileStore.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var EventEmitter = require('events').EventEmitter;
 var Dispatcher = require('../distpatcher.js');
-var EVENTS = require('../constatns/events.js');
+var EVENTS = require('./events.js');
 var DicManager = require('../dicManager.js');
 
 let _files = [];
